requirements.txt
# langchain
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-community==0.3.29
langchain-experimental==0.3.4
langchain-text-splitters>=0.3.9
langchain-anthropic==0.3.15
langgraph==0.6.6


# UI Framework
streamlit>=1.28.0
streamlit-authenticator
PyYAML
python-dotenv

# Core Dependencies
tiktoken==0.11.0
faiss-cpu==1.12.0

# Search feature
tavily-python==0.7.1
langchain-tavily==0.2.11

# Other tools
pypdf==6.0.0

# MCP (Model Context Protocol) Dependencies
langchain-mcp-adapters==0.1.9

requests>=2.31.0

# Assignment 2: Research agent tools
wikipedia>=1.4.0
arxiv==2.2.0

# Assignment 3: Multi-format document support
python-docx>=0.8.11
